+++ src/web/mod.rs
@@ -1,5 +1,5 @@
 use self::error::WebResult;
-use crate::{context::AppContext, CONFIG};
+use crate::context::AppContext;
 use axum::http::{header, HeaderValue, Method};
 use tokio::net::TcpListener;
 use tower_cookies::CookieManagerLayer;
@@ -15,15 +15,15 @@ mod view;
 pub async fn serve(ctx: AppContext) -> WebResult<()> {
     tracing::info!(
         "Starting web server on {}:{}",
-        CONFIG.web.host,
-        CONFIG.web.port
+        ctx.config.web.host,
+        ctx.config.web.port
     );
 
     let cors = CorsLayer::new()
         .allow_methods([Method::GET, Method::POST, Method::PATCH, Method::DELETE])
         .allow_headers([header::AUTHORIZATION, header::ACCEPT, header::CONTENT_TYPE])
         .allow_origin(
-            CONFIG
+            ctx.config
                 .web
                 .allowed_origins
                 .iter()
@@ -38,7 +38,7 @@ pub async fn serve(ctx: AppContext) -> WebResult<()> {
         .layer(CookieManagerLayer::new());
 
     axum::serve(
-        TcpListener::bind(format!("{}:{}", CONFIG.web.host, CONFIG.web.port)).await?,
+        TcpListener::bind(format!("{}:{}", ctx.config.web.host, ctx.config.web.port)).await?,
         app.into_make_service(),
     )
     .await?;