+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Nothing yet.
+### Changed
+
+- Removed `lazy_static` in favor of keeping config in the context (like it was before)
 
 ## [0.11.0] - 2023-12-08